Split test data in different files
