+++ test/data/confirmations/contract-interaction.ts
@@ -161,43 +161,3 @@ export const genUnapprovedContractInteractionConfirmation = ({
     userFeeLevel: 'medium',
     verifiedOnBlockchain: false,
   } as SignatureRequestType);
-
-export const genUnapprovedApproveConfirmation = ({
-  address = CONTRACT_INTERACTION_SENDER_ADDRESS,
-  chainId = CHAIN_ID,
-}: {
-  address?: Hex;
-  chainId?: string;
-} = {}) => ({
-  ...genUnapprovedContractInteractionConfirmation({ chainId }),
-  txParams: {
-    from: address,
-    data: '0x095ea7b30000000000000000000000002e0d7e8c45221fca00d74a3609a0f7097035d09b0000000000000000000000000000000000000000000000000000000000000001',
-    gas: '0x16a92',
-    to: '0x076146c765189d51be3160a2140cf80bfc73ad68',
-    value: '0x0',
-    maxFeePerGas: '0x5b06b0c0d',
-    maxPriorityFeePerGas: '0x59682f00',
-  },
-  type: TransactionType.tokenMethodApprove,
-});
-
-export const genUnapprovedSetApprovalForAllConfirmation = ({
-  address = CONTRACT_INTERACTION_SENDER_ADDRESS,
-  chainId = CHAIN_ID,
-}: {
-  address?: Hex;
-  chainId?: string;
-} = {}) => ({
-  ...genUnapprovedContractInteractionConfirmation({ chainId }),
-  txParams: {
-    from: address,
-    data: '0x095ea7b30000000000000000000000002e0d7e8c45221fca00d74a3609a0f7097035d09b0000000000000000000000000000000000000000000000000000000000000001',
-    gas: '0x16a92',
-    to: '0x076146c765189d51be3160a2140cf80bfc73ad68',
-    value: '0x0',
-    maxFeePerGas: '0x5b06b0c0d',
-    maxPriorityFeePerGas: '0x59682f00',
-  },
-  type: TransactionType.tokenMethodSetApprovalForAll,
-});
